requirements.txt
Flask==3.1.0
google-generativeai==0.8.5
python-dotenv==1.0.1
gunicorn==23.0.0
